Feature/json adapter init

+++ Engine/Render/Private/Renderer.cpp
@@ -186,7 +186,7 @@ void URenderer::GatherRenderableObjects()
 void URenderer::Render()
 {
 	//
-	// 여기에 카메라 VP 업데이트 한 번 싹 
+	// 여기에 카메라 VP 업데이트 한 번 싹
 	//
 
 	for (auto& PrimitiveComponent : PrimitiveComponents)
@@ -207,7 +207,7 @@ void URenderer::Render()
 			PrimitiveComponent->GetRelativeScale3D() );
 
 		Pipeline->SetVertexBuffer(PrimitiveComponent->GetVertexBuffer(), Stride);
-		Pipeline->Draw(PrimitiveComponent->GetVerticesData()->size(), 0);
+		Pipeline->Draw(static_cast<UINT>(PrimitiveComponent->GetVerticesData()->size()), 0);
 	}
 }
